# service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  metadataController.AddVaccineRequest:
    properties:
      note:
        example: สำหรับเด็กอายุ 6 เดือนขึ้นไป
        type: string
      vaccineName:
        example: วัคซีนไข้หวัดใหญ่
        type: string
    required:
    - vaccineName
    type: object
...
  metadataController.UpdateVaccineRequest:
    properties:
      note:
        example: สำหรับเด็กอายุ 3 เดือนขึ้นไป
        type: string
      vaccineName:
        example: วัคซีนไข้หวัดเล็ก
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Backend API for MimiCare Project (Gin + MongoDB)
  title: MimiCare API
  version: "1.0"
paths:
...
  /metaVaccine/add:
    post:
      consumes:
      - application/json
      description: ใช้สำหรับเพิ่มข้อมูลวัคซีน (admin เท่านั้น)
      parameters:
      - description: ข้อมูลวัคซีน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metadataController.AddVaccineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: เพิ่มวัคซีนใหม่
      tags:
      - MetaVaccine
  /metaVaccine/delete/{id}:
    delete:
      description: ลบข้อมูลวัคซีน (admin เท่านั้น)
      parameters:
      - description: ObjectID ของวัคซีนที่ต้องการลบ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ลบวัคซีนตาม ID
      tags:
      - MetaVaccine
  /metaVaccine/get-all:
    get:
      description: ใช้สำหรับดึงข้อมูลวัคซีนทั้งหมด (admin เท่านั้น)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลวัคซีนทั้งหมด
      tags:
      - MetaVaccine
  /metaVaccine/get/{id}:
    get:
      description: ใช้สำหรับดึงข้อมูลวัคซีนตาม ID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของวัคซีน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลวัคซีนตาม ID
      tags:
      - MetaVaccine
  /metaVaccine/update/{id}:
    put:
      consumes:
      - application/json
      description: ใช้สำหรับอัปเดตข้อมูลวัคซีนตาม ID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของวัคซีนที่ต้องการอัปเดต
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลวัคซีนที่ต้องการอัปเดต
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metadataController.UpdateVaccineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: อัปเดตข้อมูลวัคซีนตาม ID
      tags:
      - MetaVaccine
...
schemes:
- http
swagger: "2.0"